requirements.txt
Django==4.1.3
django_filter==22.1
djangorestframework==3.14.0
